All static loots + lootCount dynamic loots
This version is a mix of the other ones. It adds a random instance of
every static loots so that all containers are lootable. Then it takes
lootCount dynamic loots at random.

Note: since all statics are added by default, the server.config lootCount
for each map may have to be reduced.
Note: all containers are lootable. Their content is random (can be empty)
Note: not all dynamic loots will spawn. If they do is random, what they
are is also random.